# Remove remnants of the trace inputs struct.