--- a/tests/c/switch_non_default.c
+++ b/tests/c/switch_non_default.c
@@ -9,7 +9,7 @@
 //     yk-jit-event: stop-tracing
 //     --- Begin aot ---
 //     ...
-//     switch %{{10_1}}, bb{{bb14}}, [300 -> bb{{bb11}}, 299 -> bb{{bb12}}] [safepoint: {{safepoint_id}}, (%{{0_0}}, %{{0_1}}, %{{0_4}}, %{{0_5}}, %{{0_6}}, %{{10_1}})]
+//     switch %{{10_1}}, bb{{bb14}}, [300 -> bb{{bb11}}, 299 -> bb{{bb12}}] [safepoint: {{safepoint_id}}, (%{{0_0}}, %{{0_1}}, %{{0_4}}, %{{0_5}}, %{{0_6}})]
 //     ...
 //     --- End aot ---
 //     --- Begin jit-pre-opt ---