… tests/c/aot_debuginfo.c
… tests/c/simple_interp_loop2.c
… tests/c/switch_default.c