Even thought the trace inputs struct was no longer used, it still existed and was passed into the control point. This commit removes any remnants of the former trace inputs struct and adjusts the code (in particular the inline assembly) accordingly.

Depends on: https://github.com/ykjit/ykllvm/pull/189