@ptersilie 2024-07-30
I've measured the speed up, which is between 1.02x and 1.6x:

| test  | before | after | speedup |
-----
| `db.lua` | 194 | 185 | 1.05 |
| `bigloop.lua` |4.31 | 4.22  | 1.02 |
| `gc.lua` |525| 485| 1.08 |
| `sort.lua` | 1.53| 955| 1.6 |