@ltratt 2024-07-30
Deletes code, speeds things up -- excellent!