@ltratt 2024-07-30
Huh though this row makes no esnse:

```
sort.lua 	1.53 	955 	1.6
```

Indeed, the units in the table overall seem weird? Is it seconds or milliseconds or ... ?