@ltratt 2024-07-30
OK, I get it. Next time could you use the same units for each cell in each column? It'll stop my mind exploding!