@ptersilie 2024-07-30
I just pasted this from my vim window where I crudely pasted the numbers that `time` gave me. :laughing: I'll try to format them better next time.