@ptersilie 2024-07-30
Added clippy suggestion. Ok to squash?