@ltratt 2024-07-30
This diff is needed for `--release` mode:

```
- //     switch %{{9_1}}, bb{{bb14}}, [299 -> bb{{bb10}}, 298 -> bb{{bb12}}] [safepoint: 0i64, (%{{0_0}}, %{{0_1}}, %{{0_3}}, %{{0_4}}, %{{0_5}})]
+ //     switch %{{9_1}}, bb{{bb14}}, [299 -> bb{{bb10}}, 298 -> bb{{bb12}}] [safepoint: 1i64, (%{{0_0}}, %{{0_1}}, %{{0_3}}, %{{0_4}}, %{{0_5}})]
```

i.e. changing `0i64` to `1i64` but... why? In non-release mode `0i64` is correct. I'm not sure I know why the value has changed.