@ptersilie 2024-07-30
Which test is this? If I run `cargo test --release` I can't reproduce this.