@ptersilie 2024-07-30
Hah! Our stackmaps insertion pass is non-deterministic: https://github.com/ykjit/ykllvm/blob/main/llvm/lib/Transforms/Yk/StackMaps.cpp#L50

I'm surprised we haven't noticed this sooner. Anyway, the order of the IDs doesn't really matter, so we have two choices:
  1. Change the `std::map` to the C++ equivalent of a `IndexMap`.
  2. Make sure tests don't try to pattern match the stackmap ID.