@ltratt 2024-07-30
I presume this is the same thing Jake ran into? It sounds to me like (1) is the right thing to do, if nothing else because it'll make debugging easier. But for this PR, right now, we can do (2) while we wait for (1) to land in ykllvm.