@ptersilie 2024-07-30
I didn't know Jake ran into something similar. I agree (1) would be the right fix so we can't run into this by accident any more. Will have to look if there's an `IndexMap` in C++.

I pushed a temporary fix to get this PR in. Ok to squash?