@ltratt 2024-07-30
Please squash.

[We can easily make a "custom" indexmap, because it's basically a combo of `HashMap<T, usize>` where the `usize` is an index into a `Vec`.]